test(server): prove oversized uploads reject before the whole body is read
The existing `rejects_an_oversized_file...` test builds its body with
`Body::from(Vec<u8>)`, which hands multer a single always-ready frame.
multer drains that frame in full before `extract_file` ever observes a
chunk boundary, so the test passes even against a whole-body-buffer-then-
check implementation. It proves a reject happens, not that it happens
early, which is the whole point of the chunked read.

Add `streaming_multipart`, which delivers the body as a real multi-frame
stream that goes `Poll::Pending` between chunks (via `yield_now`) the way
a socket does, and tracks how many payload bytes the stream actually
yielded. The new test asserts `extract_file` returns `FileTooLarge`
*and* that it stopped pulling before the full 1.5 MB arrived. Reverting
the chunked loop to `field.bytes().await` fails this test (whole body
pulled: 1572864 of 1572864) while the single-frame smoke test still
passes, confirming it now guards the invariant.

Rename the single-frame test to `rejects_an_oversized_file` and note in
its doc comment that it is a smoke test only.

Move the `extract_file` and `.ifcZIP` test modules out of `mod.rs` into
sibling `extract_file_tests.rs` / `ifczip_tests.rs` files so the added
test does not push `mod.rs` past the 400-line module-size ratchet
(AGENTS.md split rule); `mod.rs` keeps only production code.
