+++ apps/server/src/routes/parse/mod.rs
@@ -227,156 +227,7 @@ fn unwrap_ifczip(
 }
 
 #[cfg(test)]
-mod extract_file_tests {
-    use super::*;
-    use axum::body::Body;
-    use axum::extract::FromRequest;
-    use axum::http::{header, Request};
-
-    const BOUNDARY: &str = "extractfileboundary";
-
-    /// Build an `axum::extract::Multipart` carrying a single `file` field with
-    /// `content`, mirroring a real POST body.
-    async fn multipart_of(content: &[u8]) -> Multipart {
-        let mut body = Vec::new();
-        body.extend_from_slice(
-            format!(
-                "--{BOUNDARY}\r\nContent-Disposition: form-data; name=\"file\"; filename=\"model.ifc\"\r\nContent-Type: application/octet-stream\r\n\r\n"
-            )
-            .as_bytes(),
-        );
-        body.extend_from_slice(content);
-        body.extend_from_slice(format!("\r\n--{BOUNDARY}--\r\n").as_bytes());
-
-        let request = Request::builder()
-            .method("POST")
-            .uri("/")
-            .header(
-                header::CONTENT_TYPE,
-                format!("multipart/form-data; boundary={BOUNDARY}"),
-            )
-            .body(Body::from(body))
-            .unwrap();
-
-        Multipart::from_request(request, &()).await.unwrap()
-    }
-
-    #[tokio::test]
-    async fn rejects_an_oversized_file_without_buffering_the_whole_body() {
-        // 1 MB ceiling; a 1.5 MB upload crosses it partway through the
-        // chunked read in `extract_file`, so this must reject before EOF
-        // rather than only after the full field has been received. Kept
-        // under axum's 2 MB default `DefaultBodyLimit` (not applied by the
-        // real router, which raises it — see `main.rs`) so this exercises
-        // `extract_file`'s own check, not the framework's raw body limit.
-        let content = vec![0u8; 1536 * 1024];
-        let mut multipart = multipart_of(&content).await;
-
-        let err = extract_file(&mut multipart, 1).await.unwrap_err();
-        assert!(matches!(err, ApiError::FileTooLarge { max_mb: 1 }));
-    }
-
-    #[tokio::test]
-    async fn accepts_a_file_within_the_ceiling() {
-        // Also exercises the non-gzip/non-zip branch under the new
-        // chunk-by-chunk read (previously a single `Field::bytes()` call).
-        let content = vec![0u8; 512 * 1024];
-        let mut multipart = multipart_of(&content).await;
-
-        let bytes = extract_file(&mut multipart, 1).await.unwrap();
-        assert_eq!(bytes.len(), content.len());
-    }
-}
+mod extract_file_tests;
 
 #[cfg(test)]
-mod ifczip_tests {
-    use super::*;
-    use std::io::Write;
-    use zip::write::{SimpleFileOptions, ZipWriter};
-
-    const STEP: &str = "ISO-10303-21;\nHEADER;\nENDSEC;\nDATA;\nENDSEC;\nEND-ISO-10303-21;";
-    // Deliberately larger than the raw/gzip max the server would apply, so
-    // callers can pass a real ceiling.
-    const BIG: usize = 512 * 1024 * 1024;
-
-    /// Build an in-memory zip from `(name, content)` pairs (Stored so declared
-    /// uncompressed sizes are exact for the zip-bomb test).
-    fn make_zip(entries: &[(&str, &str)]) -> Vec<u8> {
-        let mut buf = Cursor::new(Vec::new());
-        {
-            let mut zip = ZipWriter::new(&mut buf);
-            let opts =
-                SimpleFileOptions::default().compression_method(zip::CompressionMethod::Stored);
-            for (name, content) in entries {
-                zip.start_file(*name, opts).unwrap();
-                zip.write_all(content.as_bytes()).unwrap();
-            }
-            zip.finish().unwrap();
-        }
-        buf.into_inner()
-    }
-
-    #[test]
-    fn extracts_the_single_model_entry() {
-        let zip = make_zip(&[("model.ifc", STEP)]);
-        let out = unwrap_ifczip(&zip, BIG, 512).unwrap();
-        assert_eq!(String::from_utf8(out.to_vec()).unwrap(), STEP);
-    }
-
-    #[test]
-    fn matches_ifcxml_case_insensitively_from_a_nested_path() {
-        let zip = make_zip(&[("nested/dir/Model.IFCXML", "<ifcXML/>")]);
-        let out = unwrap_ifczip(&zip, BIG, 512).unwrap();
-        assert_eq!(String::from_utf8(out.to_vec()).unwrap(), "<ifcXML/>");
-    }
-
-    #[test]
-    fn ignores_referenced_resources_alongside_the_model() {
-        let zip = make_zip(&[("model.ifc", STEP), ("resources/texture.png", "not-a-png")]);
-        let out = unwrap_ifczip(&zip, BIG, 512).unwrap();
-        assert_eq!(String::from_utf8(out.to_vec()).unwrap(), STEP);
-    }
-
-    #[test]
-    fn rejects_an_archive_with_no_model_entry() {
-        let zip = make_zip(&[("readme.txt", "hello")]);
-        let err = unwrap_ifczip(&zip, BIG, 512).unwrap_err();
-        assert!(matches!(err, ApiError::BadRequest(m) if m.contains("no .ifc/.ifcxml entry")));
-    }
-
-    #[test]
-    fn rejects_an_archive_with_multiple_model_entries() {
-        let zip = make_zip(&[("a.ifc", STEP), ("b.ifc", STEP)]);
-        let err = unwrap_ifczip(&zip, BIG, 512).unwrap_err();
-        assert!(matches!(err, ApiError::BadRequest(m) if m.contains("expected exactly one")));
-    }
-
-    #[test]
-    fn rejects_an_entry_over_the_size_ceiling() {
-        // model.ifc is ~60 bytes; a 10-byte ceiling trips the zip-bomb guard
-        // on the declared uncompressed size before decompressing.
-        let zip = make_zip(&[("model.ifc", STEP)]);
-        let err = unwrap_ifczip(&zip, 10, 1).unwrap_err();
-        assert!(matches!(err, ApiError::FileTooLarge { max_mb: 1 }));
-    }
-
-    #[test]
-    fn extracts_a_deflate_compressed_model_entry() {
-        // Real buildingSMART .ifcZIP containers are DEFLATE-compressed, not
-        // Stored. This exercises the actual `deflate` feature path so a
-        // mis-wired Cargo.toml feature fails here instead of only in production
-        // (UnsupportedArchive at decode time).
-        let mut buf = Cursor::new(Vec::new());
-        {
-            let mut zip = ZipWriter::new(&mut buf);
-            let opts =
-                SimpleFileOptions::default().compression_method(zip::CompressionMethod::Deflated);
-            zip.start_file("model.ifc", opts).unwrap();
-            zip.write_all(STEP.as_bytes()).unwrap();
-            zip.finish().unwrap();
-        }
-        let zip = buf.into_inner();
-        let out = unwrap_ifczip(&zip, BIG, 512).unwrap();
-        assert_eq!(String::from_utf8(out.to_vec()).unwrap(), STEP);
-    }
-}
+mod ifczip_tests;
